pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "PROJECT"
version = "0.1.0"
description = "Telegram sample bot adapter"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=22.7,<23",
    "python-dotenv>=1.2.2,<2",
    "httpx>=0.28,<0.29",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
